requirements.txt
curl_cffi>=0.7.0
kiteconnect>=4.2.0
python-binance>=1.0.19
yfinance>=0.2.38
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
smartapi-python>=1.3.4
pyotp>=2.9.0
feedparser>=6.0.0
anthropic>=0.25.0